scheduler/src/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-debug-toolbar==4.3.0
django-split-settings==1.2.0
flake8==6.1.0
gunicorn==21.2.0
uvicorn==0.29.0
mccabe==0.7.0
memory-profiler==0.61.0
mypy==1.8.0
mypy-extensions==1.0.0
packaging==23.2
psutil==5.9.8
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyflakes==3.1.0
python-dotenv==1.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
uWSGI==2.0.24
requests==2.32.3
fastapi-limiter==0.1.6
django-celery-beat
